Reapply "feat hard work..."
This reverts commit bac254bc2d2f575907a78d12b7368b7706ac0082.

+++ Domain/Entities/DesignCategory.cs
@@ -7,7 +7,7 @@ namespace Domain.Entities
     {
         [Key]
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
-        public int DesignCategoryId { get; set; }
+        public long DesignCategoryId { get; set; }
         public string DesignCategoryName { get; set; }
         public List<Design> Designs { get; set; }
     }
